requirements.txt
streamlit>=1.35.0
pydantic>=2.7.0
numpy>=1.26.0
pandas>=2.2.0
PyMuPDF>=1.24.0
pdfplumber>=0.11.0
Pillow>=10.3.0
faiss-cpu>=1.8.0
transformers>=4.41.0
torch>=2.2.0
tqdm>=4.66.0
requests>=2.32.0
python-dotenv>=1.0.1
openai>=1.35.0
